Refactor AdminProgramController.NewProgram form class to forms.ProgramForm.
