+++ universal-application-tool-0.0.1/app/forms/ProgramForm.java
@@ -1,4 +1,22 @@
 package forms;
 
 public class ProgramForm {
+  private String name;
+  private String description;
+
+  public String getName() {
+    return name;
+  }
+
+  public void setName(String name) {
+    this.name = name;
+  }
+
+  public String getDescription() {
+    return description;
+  }
+
+  public void setDescription(String description) {
+    this.description = description;
+  }
 }
